requirements.txt
absl-py==0.9.0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.1
contextlib2==0.6.0.post1
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
httplib2==0.9.2
idna==2.6
importlib-metadata==1.6.0
importlib-resources==1.4.0
injector==0.18.3
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
oauth==1.0.1
object-detection==0.1
olefile==0.45.1
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
peewee==3.13.2
pexpect==4.2.1
pickleshare==0.7.5
picsellia==0.0.8
Pillow==5.1.0
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
protobuf-to-dict==0.1.0
ptyprocess==0.6.0
py3nvml==0.2.6
pycairo==1.16.2
pycocotools==2.0.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.14.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.18
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xkit==0.0.0
xmltodict==0.12.0
zipp==3.1.0
zope.interface==4.3.2
